Time: 0 ms (100.00%) | Memory: 30.2 MB (85.64%) - LeetSync

+++ 4284-smallest-stable-index-i/smallest-stable-index-i.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int firstStableIndex(vector<int>& nums, int k) {
-    int n =nums.size();
+    long long n =nums.size();
     for(int i =0;i<n;i++){
         int maxint = INT_MIN;
         for(int j =0;j<=i;j++){
